# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/deployments": {
      "post": {
        "operationId": "DeploymentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeploymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "deployments"
        ]
      },
      "get": {
        "operationId": "DeploymentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeploymentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "deployments"
        ]
      }
    },
...
    "/deployments/{id}/env": {
      "get": {
        "operationId": "EnvVarsController_list",
        "parameters": [
          {
            "name": "service",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaskedEnvVarDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "env"
        ]
      }
    },
    "/deployments/{id}/env/{key}": {
      "put": {
        "operationId": "EnvVarsController_set",
        "parameters": [
          {
            "name": "service",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEnvVarDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "env"
        ]
      },
      "patch": {
        "operationId": "EnvVarsController_updateMeta",
        "parameters": [
          {
            "name": "service",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEnvVarMetaDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "env"
        ]
      },
      "delete": {
        "operationId": "EnvVarsController_remove",
        "parameters": [
          {
            "name": "service",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "env"
        ]
      }
    },
...
    "/backups/{id}": {
      "delete": {
        "operationId": "BackupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/schedules": {
      "get": {
        "operationId": "BackupSchedulesController_list",
        "parameters": [
          {
            "name": "deploymentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupScheduleDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      },
      "post": {
        "operationId": "BackupSchedulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBackupScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/schedules/{id}": {
      "patch": {
        "operationId": "BackupSchedulesController_setEnabled",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBackupScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      },
      "delete": {
        "operationId": "BackupSchedulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
...
    "/deployments/{id}/containers": {
      "get": {
        "operationId": "DeploymentVolumesController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContainerDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "deployments"
        ]
      }
    },
    "/deployments/{id}/volumes/{name}/reset": {
      "post": {
        "operationId": "DeploymentVolumesController_reset",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "deployments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Willy API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "CreateDeploymentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,40}$",
            "example": "my-app"
          },
          "type": {
            "type": "string",
            "enum": [
              "WEB",
              "WORKER",
              "CRON"
            ]
          },
          "gitUrl": {
            "type": "string",
            "example": "https://github.com/owner/repo.git"
          },
          "gitRef": {
            "type": "string",
            "example": "main"
          },
          "imageRef": {
            "type": "string",
            "example": "nginx:1.27",
            "description": "For the IMAGE strategy."
          },
          "buildStrategy": {
            "type": "string",
            "enum": [
              "DOCKERFILE",
              "COMPOSE",
              "IMAGE"
            ]
          },
          "dockerfilePath": {
            "type": "string"
          },
          "composeFilePath": {
            "type": "string",
            "example": "docker-compose.yml"
          },
          "composeWebService": {
            "type": "string",
            "description": "Compose service to route + monitor."
          },
          "healthCheckPath": {
            "type": "string",
            "example": "/"
          },
          "runCommand": {
            "type": "string"
          },
          "cronExpr": {
            "type": "string",
            "example": "0 3 * * *"
          },
          "domain": {
            "type": "string",
            "example": "app.example.com"
          },
          "domainPort": {
            "type": "number",
            "minimum": 1,
            "maximum": 65535,
            "description": "Port the primary domain routes to (defaults to the image's first exposed port)."
          },
          "domainService": {
            "type": "string",
            "description": "Compose service the primary domain routes to (compose deployments)."
          },
          "gitToken": {
            "type": "string",
            "description": "Personal access token for a private repo (encrypted at rest)."
          },
          "memoryLimitMb": {
            "type": "number",
            "minimum": 16
          },
          "nanoCpus": {
            "type": "number",
            "description": "CPU limit in nano-CPUs (1 CPU = 1e9)"
          },
          "capAdd": {
            "example": [
              "NET_ADMIN"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capDrop": {
            "example": [
              "ALL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "StrategyConfigDto": {
        "type": "object",
        "properties": {
          "dockerfilePath": {
            "type": "string"
          },
          "composeFilePath": {
            "type": "string"
          },
          "composeWebService": {
            "type": "string"
          },
          "imageRef": {
            "type": "string"
          }
        }
      },
      "HealthcheckDto": {
        "type": "object",
        "properties": {
          "test": {
            "type": "string",
            "example": "curl -f http://localhost/health"
          },
          "interval": {
            "type": "string",
            "example": "30s"
          },
          "timeout": {
            "type": "string",
            "example": "10s"
          },
          "retries": {
            "type": "number",
            "example": 3
          },
          "startPeriod": {
            "type": "string",
            "example": "5s"
          }
        },
        "required": [
          "test"
        ]
      },
      "DeploymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "WEB",
              "WORKER",
              "CRON"
            ]
          },
          "gitUrl": {
            "type": "string"
          },
          "gitRef": {
            "type": "string"
          },
          "buildStrategy": {
            "type": "string",
            "enum": [
              "DOCKERFILE",
              "COMPOSE",
              "IMAGE"
            ]
          },
          "strategyConfig": {
            "$ref": "#/components/schemas/StrategyConfigDto"
          },
          "runCommand": {
            "type": "string",
            "nullable": true
          },
          "cronExpr": {
            "type": "string",
            "nullable": true
          },
          "webServicePort": {
            "type": "number",
            "nullable": true
          },
          "healthCheckPath": {
            "type": "string"
          },
          "autoDeploy": {
            "type": "boolean"
          },
          "restartPolicy": {
            "type": "string",
            "enum": [
              "NO",
              "ON_FAILURE",
              "ALWAYS",
              "UNLESS_STOPPED"
            ]
          },
          "memoryLimitMb": {
            "type": "number",
            "nullable": true
          },
          "nanoCpus": {
            "type": "number",
            "nullable": true
          },
          "capAdd": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capDrop": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logMaxSizeMb": {
            "type": "number",
            "nullable": true
          },
          "logMaxFiles": {
            "type": "number",
            "nullable": true
          },
          "healthcheck": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/HealthcheckDto"
              }
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "CREATED",
              "DEPLOYING",
              "RUNNING",
              "DEGRADED",
              "STOPPED",
              "ERROR"
            ]
          },
          "activeReleaseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "gitCredentialId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "primaryDomain": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "gitUrl",
          "gitRef",
          "buildStrategy",
          "strategyConfig",
          "runCommand",
          "cronExpr",
          "webServicePort",
          "healthCheckPath",
          "autoDeploy",
          "restartPolicy",
          "memoryLimitMb",
          "nanoCpus",
          "capAdd",
          "capDrop",
          "logMaxSizeMb",
          "logMaxFiles",
          "healthcheck",
          "state",
          "activeReleaseId",
          "gitCredentialId",
          "createdAt",
          "updatedAt",
          "primaryDomain"
        ]
      },
...
      "MaskedEnvVarDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "BUILD",
              "RUNTIME",
              "BOTH"
            ]
          },
          "isSecret": {
            "type": "boolean"
          },
          "value": {
            "type": "string",
            "nullable": true,
            "description": "Plaintext value, or null for secrets."
          }
        },
        "required": [
          "key",
          "scope",
          "isSecret",
          "value"
        ]
      },
      "SetEnvVarDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "BUILD",
              "RUNTIME",
              "BOTH"
            ]
          },
          "isSecret": {
            "type": "boolean"
          }
        },
        "required": [
          "value"
        ]
      },
      "UpdateEnvVarMetaDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "BUILD",
              "RUNTIME",
              "BOTH"
            ]
          },
          "isSecret": {
            "type": "boolean"
          }
        }
      },
...
      "BackupScheduleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deploymentId": {
            "type": "string",
            "nullable": true
          },
          "target": {
            "type": "string"
          },
          "cron": {
            "type": "string"
          },
          "retention": {
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          },
          "lastRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "deploymentId",
          "target",
          "cron",
          "retention",
          "enabled",
          "lastRunAt",
          "createdAt"
        ]
      },
      "CreateBackupScheduleDto": {
        "type": "object",
        "properties": {
          "target": {
            "type": "string",
            "example": "willy_blog_data"
          },
          "cron": {
            "type": "string",
            "example": "0 3 * * *"
          },
          "retention": {
            "type": "number",
            "example": 7
          },
          "deploymentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "target",
          "cron"
        ]
      },
      "UpdateBackupScheduleDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ]
      },
...
      "VolumeMountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "rw": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "destination",
          "rw"
        ]
      },
      "NetworkInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ip": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "ip"
        ]
      },
      "DeclaredHealthcheckDto": {
        "type": "object",
        "properties": {
          "test": {
            "example": [
              "CMD-SHELL",
              "curl -f http://localhost/health"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "interval": {
            "type": "string",
            "nullable": true,
            "example": "30s"
          },
          "timeout": {
            "type": "string",
            "nullable": true,
            "example": "10s"
          },
          "retries": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "startPeriod": {
            "type": "string",
            "nullable": true,
            "example": "5s"
          }
        },
        "required": [
          "test",
          "interval",
          "timeout",
          "retries",
          "startPeriod"
        ]
      },
      "ContainerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "running": {
            "type": "boolean"
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VolumeMountDto"
            }
          },
          "service": {
            "type": "string",
            "nullable": true
          },
          "networks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NetworkInfoDto"
            }
          },
          "exposedPorts": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "health": {
            "type": "string",
            "nullable": true,
            "example": "healthy"
          },
          "declaredHealthcheck": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeclaredHealthcheckDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "running",
          "volumes",
          "service",
          "networks",
          "exposedPorts",
          "health",
          "declaredHealthcheck"
        ]
      },
...
    }
  }
}